refactor(http): move ErrUnauthorized to http package
