+++ api/http/errors/errors.go
@@ -1,0 +1,8 @@
+package errors
+
+import "errors"
+
+// General errors.
+var (
+	ErrUnauthorized = errors.New("Unauthorized")
+)
